requirements.txt
fastapi
uvicorn
pandas
requests
python-dotenv
python-multipart
pydantic
httpx
typing
sqlalchemy[asyncio]
asyncpg
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
